refactor(form): rename the otp folder to its node suffix; sort package indexes
components/otp-input becomes components/otp with OtpAdapter, matching
the field.otp node like every other folder; the ui and form package
indexes are sorted by module path.
