+++ packages/form/resources/js/index.ts
@@ -1,9 +1,13 @@
-export * from "./components/index";
-export { FormFieldFrame, type FormFieldControlProps } from "./components/base/field";
-export { FormValuesProvider } from "./hooks/values";
 export { ActionForm, useLazyActionForm } from "./action-form";
+export { FormFieldFrame, type FormFieldControlProps } from "./components/base/field";
+export { Checkbox } from "./components/checkbox/checkbox";
+export * from "./components/color-picker/color-picker";
+export * from "./components/index";
+export * from "./components/otp/otp";
+export { default as PasswordInput } from "./components/password-input/password-input";
+export { Textarea } from "./components/textarea/textarea";
+export { FormValuesProvider } from "./hooks/values";
 export { formComponents } from "./plugin";
-export type * from "./types";
 export * from "./primitives/affix-group";
 export * from "./primitives/combobox";
 export * from "./primitives/date-picker";
@@ -13,8 +17,4 @@ export { default as InputError } from "./primitives/input-error";
 export * from "./primitives/label";
 export * from "./primitives/option-cards";
 export * from "./primitives/time-picker";
-export { Checkbox } from "./components/checkbox/checkbox";
-export * from "./components/color-picker/color-picker";
-export * from "./components/otp-input/otp-input";
-export { default as PasswordInput } from "./components/password-input/password-input";
-export { Textarea } from "./components/textarea/textarea";
+export type * from "./types";
